+++ src/kiara/module.py
@@ -23,7 +23,7 @@ from kiara.metadata.module_models import (
     KiaraModuleTypeMetadata,
 )
 from kiara.module_config import KIARA_CONFIG, ModuleTypeConfig
-from kiara.operations import OperationConfig
+from kiara.operations import Operation
 from kiara.processing import JobLog
 from kiara.utils import StringYAML, is_debug
 from kiara.utils.modules import create_schemas, overlay_constants_and_defaults
@@ -240,9 +240,7 @@ class KiaraModule(typing.Generic[KIARA_CONFIG], abc.ABC):
     @classmethod
     def retrieve_module_profiles(
         cls, kiara: "Kiara"
-    ) -> typing.Mapping[
-        str, typing.Union[typing.Mapping[str, typing.Any], OperationConfig]
-    ]:
+    ) -> typing.Mapping[str, typing.Union[typing.Mapping[str, typing.Any], Operation]]:
         pass
 
     @classmethod