+++ src/kiara/utils/class_loading.py
@@ -20,7 +20,7 @@ from kiara.utils import (
 
 if typing.TYPE_CHECKING:
     from kiara.module import KiaraModule
-    from kiara.operations import Operations
+    from kiara.operations import OperationType
 
 log = logging.getLogger("kiara")
 
@@ -276,13 +276,13 @@ def find_all_value_types() -> typing.Dict[str, typing.Type["ValueType"]]:
     return all_value_types
 
 
-def find_all_operation_types() -> typing.Dict[str, typing.Type["Operations"]]:
+def find_all_operation_types() -> typing.Dict[str, typing.Type["OperationType"]]:
 
-    from kiara.operations import Operations
+    from kiara.operations import OperationType
 
     return load_all_subclasses_for_entry_point(
         entry_point_name="kiara.operation_types",
-        base_class=Operations,
+        base_class=OperationType,
         set_id_attribute="_operation_type_name",
         remove_namespace_tokens=["core."],
     )
@@ -388,12 +388,12 @@ def find_value_types_under(
 
 def find_operations_under(
     module: typing.Union[str, ModuleType], prefix: typing.Optional[str] = ""
-) -> typing.Mapping[str, typing.Type["Operations"]]:
+) -> typing.Mapping[str, typing.Type["OperationType"]]:
 
-    from kiara.operations import Operations
+    from kiara.operations import OperationType
 
     return find_subclasses_under(
-        base_class=Operations,
+        base_class=OperationType,
         module=module,
         prefix=prefix,
         remove_namespace_tokens=[],